chore: relocate the model-catalog update tooling to @qvac/inference
