+++ packages/inference/src/models/update-models/types.ts
@@ -1,4 +1,4 @@
-import type { ModelRegistryEntryAddon, ModelRegistryEngine } from '../../schemas/registry'
+import type { ModelRegistryEntryAddon, ModelRegistryEngine } from '../../surface'
 
 export interface ShardInfo {
   isSharded: true
