+++ packages/inference/test/companion-detection.test.ts
